requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery
redis
python-dotenv
drf-spectacular